Error handling
fixed a glitch where if the jump and crawl keys were pressed too quickly, the cat would be stuck crouching, but unable to uncrouch.
implemented the forwards movement. so, now the cat will always be moving forward(or in this case to the right due to perspective)
changed the movement so that switching lanes is a function of its own rather than update